refactor(electron/player): decompose usePlayerMachine into focused sub-hooks — Phase 3.5
333 LOC god-hook → 33 LOC composition + 6 focused sub-hooks under hooks/player/:
- usePlayerActor (89)     : actor lifecycle (machine.provide, INITIALIZE, CLEANUP, send)
- useMachineToStoreSync (66): snapshot → playerStore + visual-cue notify
- useParagraphSubscriptions (79): three zustand→machine subs + TTS prefetch
- useNavBridge (40)       : navStore.navState leaving idle → PAGE_NAVIGATING
- usePdfSeed (40)         : pdfStore seed-on-mount path
- useResumeWrite (68)     : startResumeWriteSubscription wrapper (re-export preserved)

usePlayerMachine.ts now meets the Phase 3.5 ≤100 LOC exit criterion.
1464 + 511 tests green.
